switch Create pages to new versions
