+++ BlazorWebApp/Pages/CreateRTBase.cs
@@ -13,94 +13,63 @@ namespace BlazorWebApp.Pages
 
     public class CreateRTBase : ComponentBase
     {
-        protected IBrowserFile file = null!;
+        protected IBrowserFile? file = null!;
 
-        protected RTModalComponent? childmodal { get; set; } = null!;
-        protected string? DebugMsg { get; set; } = "API Return Message";
+        protected OneItem? childOneItem { get; set; } = null!;
+    
+        
         protected string apiBase { get; set; } = "http://127.0.0.1:5026";
 
         protected string apiRoute { get; set; } = "/RememberThis/rtMulti";
 
         public string apiUrl { get; set; } = string.Empty;
         protected bool ShowPopUp { get; set; } = false;
+        
 
         //pre-build item in dev mode so we don't have to type one in each time we test
-        protected rtItem thisrtItem { get; set; } =
-            new rtItem
-            {
-                rtId = 1001,
-                rtUserName = "Cosmo",
-                rtDescription = "fun time digging hole for bone",
-                rtLocation = "backyard",
-                rtDateTime = DateTime.UtcNow
-            };
+        protected rtItem? thisrtItem { get; set; } = null!;
+        
+        protected string createCardTitle { get; set; } = "Add a memory you want to save and share later!";
+        
         protected EditContext? rtItemEditContext;
-        protected string ChildModalBody { get; set; } = string.Empty;
-
+        protected string? DebugMsg { get; set; } = "API Return Message";
 
         [Inject]
         protected IHttpClientFactory ClientFactory { get; set; } = null!;
 
-        [Inject]
-        protected IJSRuntime jsRuntime { get; set; } = null!;
+   
 
         [Inject]
         protected ItemService _ItemService { get; set; } = null!;
 
-         [Inject]
-        protected IConfiguration Config { get; set; } = null!;
-
         protected override void OnInitialized()
         {
-            rtItemEditContext = new(thisrtItem);
-        }
-        protected async Task SelectedFileProcess(InputFileChangeEventArgs e)
-        {
-            long _fileSizeLimit = Config.GetValue<long>("FileSizeLimit");
-
-            if (!((e.File.Size > 0) && (e.File.Size < _fileSizeLimit)))
+            thisrtItem = new rtItem
             {
-                ChildModalBody = "File size invalid";
-                childmodal!.Open();
-                await jsRuntime.InvokeVoidAsync("ResetFilePicker");
+                rtId = 1001,
+                rtUserName = "Cosmo",
+                rtDescription = string.Empty,
+                rtLocation = string.Empty,
+                rtDateTime = DateTime.UtcNow
+            };
 
-            }
-            else
-            {
-                // file size is good so save to class variable and update preview on screen
-                file = e.File;
-                await jsRuntime.InvokeVoidAsync("loadFileJS");
-            }
-        }
+            // rtItemEditContext = new(thisrtItem);
+        }      
 
-        protected void DisplayBtnClicked(string _btnClicked)
-        {
-            DebugMsg = _btnClicked;
-
-        }
-
-        protected async Task ConfirmInternalNavigation(LocationChangingContext context)
-        {
-            var confirmed = await jsRuntime.InvokeAsync<bool>("confirm", "Discard your changes?");
-
-            if (!confirmed)
-            {
-                context.PreventNavigation();
-            }
-        }
-
-
-        protected async Task SubmitForm()
+        protected async Task SubmitForm(IBrowserFile fileFromChild)
         {
             string PersistReturnMsg = string.Empty;
+
+            file = fileFromChild;
 
             // need to add code to handle no image selected
             var ms = new MemoryStream();
             await file.OpenReadStream(1024 * 1024 * 10).CopyToAsync(ms);
             ms.Position = 0;
 
-            PersistReturnMsg = await _ItemService.AddItem(thisrtItem, ms, file.Name, file.ContentType);
+            PersistReturnMsg = await _ItemService.AddItem(thisrtItem!, ms, file.Name, file.ContentType);
 
+            
             DebugMsg = PersistReturnMsg;
