--- BlazorWebApp/Pages/CreateRTNEW.razor
@@ -1,19 +1,0 @@
-@page "/CreateRTNEW"
-@inherits CreateRTBaseNEW
-
-@parentMyStuff
-
-<OneItem    @ref="childOneItem"             
-            thisChildrtItem="thisrtItem"             
-            CallParentSubmit="SubmitForm"
-            >
-</OneItem>
-
-
-
-
-@DebugMsg
-
-
-
-
