+++ BlazorWebApp/Pages/OneItem.razor
@@ -51,7 +51,7 @@ it takes the following parameters
 
 <div class="card">
     <div class="card-header">
-        <h3 style="text-align:center">Add a memory you want to save and share later!</h3>
+        <h3 style="text-align:center">@childCardTitle</h3>
     </div>
     <div class="card-body">        
         <EditForm EditContext="@rtItemChildEditContext" OnValidSubmit="SubmitECB">
@@ -109,6 +109,9 @@ it takes the following parameters
     public IBrowserFile childFile { get; set; } = null!;
      
 
+    [Parameter]
+    public string? childCardTitle { get; set; } = null!;
+    
     [Parameter]
     public rtItem? thisChildrtItem { get; set; } = null!;
